ExpEng types: Added new Match and State types and match function (comments)
This should be the final layout of how the NFA will work.  It has tags to handle recursive printing and Eq comparisons (not yet implemented), a match for the different kinds of matching and special tagging for start and end of a group clause (the idea here is to allow the NFA to record when a grouping is finished based on seeing these tags).

Also added initial cut at match function for running the states.