Added tag for Final (comments)
The problem is, later we will allow partial matches to run actions, but what happens in a pattern like "a+b*a" matched agains a string like "aaaa" (iterate 4 'a')?  That will produce 3 successful partial matches ("aa", "aaa", "aaaa").  Now that we tag Final we can check all the matches we got in the end and when we see multiple matches with the same tag we can simply take the longest one.