Changed match to use a monad transformer (comments)
This was needed so the group handling code can store the locations that it matched at (not yet completely implemented).  The old, non-transformer version has been kept so we can do speed tests later to find out how much the monad transformer is slowing us down.  If it's dramatic we may have to unroll the transformer ourselves by hand.