NFA Matcher: added type signatures, comment and tmp fix (comments)
The comment is about the nature of the tmp fix: namely that we want the aux function to be match' not match'' but match' is currently taken by the baseline (more efficient) version of match.  We'll need that later to benchmark the speed of the engine with a transformer and without.