Added new Accept tag and handling for it (comments)
The new Accept tag is for expressions that don't have to reach the end of the string to match.  To handle this there is the Accept tag that records a win if it's seen and the Final tag remains which only wins at the end of the string.  Any wins that are seen are recorded in the state.