Matcher - bug fix: Group tags were causing match fails
Any (Haskell) patterns that don't consume any of the string have to be matched before the fail on empty string.  This is a long standing bug where, for example, a pattern that ended with a group (even potentially, as in a group followed by a star expression) it could not match because empty string fail would match before the CloseGroup match was checked.  Even if we check CloseGroup first, that could still fail because the CloseGroup could hold a Split where one of it's clauses was a Final.  Since the Split wasn't broken down, the Final state wouldn't be seen even though it matches (Split consumes no characters).