Parser: moved group grammar location (comments)
This change creates the group structure that I was producing by hand, which I think is cleaner.  It will potentially cause multiple open groups per single close group (or no close group) but I think that's fine.

NOTE: this implementation contains a bug: if you have a structure like "^a(b|x)*c$" the created NFA will put the OpenGroup in the right place, in front of the b/x split.  The problem is, this consumes no input so it always matches, causing a 1 to be logged in the state for that group even if the group match ends up failing.  So if you match the above on a string like "ac" it will incorrectly report a group match of "(1,1)".  Of course a zero size capture is impossible and could be filtered, the issue comes from the underlying problem stated above and filtering out zero size captures only hides the issue, it will pop up in other locations.

To resolve this we either have to move the OpenGroup in such a way that it is only encountered after input from its group has been consumed or (much better) find a way to let a given traversal carry it's encountered Open/CloseGroups with it, so if the traversal fails entirely, those entries would all be dropped.