Matcher: reversed order of Monads to fix bogus group matching (comments)
Ok, so previously we had (ListT (StateT)), which meant the list monad with a state that was visible to all traversals.  This is not what we actually wanted as this caused one traversal to record an OpenGroup that another traversal used even though it didn't see that specific one.  What we actually wanted was the state to be only visible to a given traversal.  So to achieve this we simply reverse the monads.  List is now the "inside" monad and State the outside.  This has the effect that each traversal has its own state for bookkeeping, so there is no longer a danger of a traversal seeing a CloseGroup and using the OpenGroup of a totally different traversal.

NOTE: This does cause one bug, however.  Since the State is now only visible to a specific traversal, and not all of them, any matches that occur before the end of the string will be forgotten since match keeps running until the string is exhausted.  To fix this, we'll have to wrap what we have now around another state monad.  This new State will receive the "Results" data which will now have to include the final group results as well as the Accept state itself.